+++ doctr/datasets/funsd.py
@@ -10,7 +10,6 @@ from pathlib import Path
 from typing import List, Dict, Any, Tuple, Optional, Callable
 import tensorflow as tf
 
-from doctr.documents.reader import read_img
 from .core import VisionDataset
 
 __all__ = ['FUNSD']
